+++ changelog.md
@@ -5,6 +5,7 @@
  - remove `.initKeyEvent()` and `.initMutationEvent()`
  - fix: `.dispatchEvent()` must return a value
  - fix: bug in `.stopPropagation()`
+ - fix: `.getComputedStyle(el).animationName` may be equal to `none, none, …` in IE11
 
 ##v0.4.1
 
@@ -36,5 +37,5 @@
 
  - new: `Set()`, `Map()`, `WeakSet()` and `WeakMap()` polyfills
  - new: [`handleEvent`](https://github.com/Octane/jsCore/issues/1) support in IE8
- - fix: `clearImmediate()` for IE8
+ - fix: `.clearImmediate()` for IE8
  - fix: `.dataset` and `.classList` polyfills use [`ActiveXObject("htmlfile")`](https://github.com/es-shims/es5-shim/issues/152) in IE8